Undo changes related to using types instead of instances of types.
I couldn't get that to work properly with dispatch.
Also, replace Bernoulli by Binomial as the distribution.  One method for `glm` now accepts explicit weights and offset as optional arguments.  To fit a binomial glm use the proportion of successes as the response and the number of trials as the weights.

Used loops and @eval for the definition of the linkfun!, linkinv! and mueta! methods.