change symbol argument in Cholesky to Char
I got an error when trying to run a lm model.  It actually make the model but fails in show because it failed in Stderr in the Coeftable which called the vcov function which called cholfact with a DensePredQR argument which called Cholesky.  The second argument to Cholesky in Base.LinAlg.Cholesky is a Char instead of a symbol:

Cholesky{T}(UL::Array{T,2},uplo::Char)

After the change it worked fine.