Add Bernoulli distribution.
Many, if not most, cases of fitting a GLM with a `Binomial()` Distributions
could be written as `Bernoulli()`.  There is a lot of messiness dealing with
`Binomial()`.  We should encourage the use of `Bernoulli()` instead.

The problem with `Binomial()` is that the distribution does not depend solely
on `μ`, the probability of success.  You also need to know `n`, the number of
trials which is passed (incorrectly) as the weights.  If `n` is always `1`,
just call the distribution `Bernoulli()`.