Evaluate inverse link and its derivative simultaneously (#192)
* Introduce inverselink generic and methods

* Spacing

* Add inverselink returning a tuple

* Correct a typo

* Add InverseSquareLink and InverseGaussian distribution

Modify tests to use approximate comparisons.
The ftest tables now fail locally, perhaps due to tab/spaces
inconsistency. I will change them to doctests.

* Clean up performance test.

* Delete trailing whitespace

* Switch docs to Documenter

jldoctest chunks are failing because GLM namespace is not visible.  Help
would be appreciated.

* Get tests working.

* Export more functions.

* Add documentation and doctests

* Add a DocTestSetup stanza

* Don't test on nightlies until StatsFuns can be compiled

* Restore testing on nightlies; add .gitignore and mkdocs.yml

* Clean up code

* Correct method signatures for `iscanonical`.

The canonical links for `Gamma` and `InverseGaussian` are flagged at
present because it is not true that the variance function is equal to
`dμ/dη`.  The variance function is a negative multiple of the
derivative.

* Move test matching `ftest` output to a doctest

* Reformulate `iscanonical` methods to work with v0.5

* Drop support for v0.5

* Spacing and use of `1 +` in formula

* Change `iscanonical` to `cancancel`.

Also use `L()` instead of `Link(r)`.

* Add tests for `cancancel` and `InverseGaussian` GLM