fixing examples
I have made the first two examples run and corrected the outputs following what my REPL prints.
In the future probably the way of getting confidence should be fixed (now it is not nice, but this is the only way I could make it work). I will prepare a PR if it is OK, but I start to use GLM.jl in teaching so it was important for me to make examples README.jl work.