Add basic documentation for `fit` and `glm`
I was unable to find any documentation on how to fit a GLM to data, beyond a few
examples in the docs. I have added a small amount of documentation for how to
use `fit`, and added them to the Documenter markdown. I have also re-iterated
that `glm` is an alias for `fit`, which wasn't clear to me when I first started
using this package.