More small improvements to CoefTable
Use the new argument in StatsBase 0.30 to print t/z values with fewer digits.
Rename the "t/z" value column to just "t/z" to avoid wasting space which isn't needed
to print the digits (removing an inconsistency with other columns which use upper case).
Also rename the "Estimate" column to "Coef.", which is more consistent with
the name of the accessor function `coef`, clearer and more correct since all columns
are actually estimates.

Update doctests output in manual.