Add geometric distribution (#480)
The geometric distribution is used in sports such as football, cricket, etc. The probability that a batsman can make a successful hit on the ball before he gets out by the bowler, is calculated efficiently using the geometric probability distribution function. Now each player will have their own distribution depending on their batting ability. So, the probability of successfully playing a ball can be modelled as a function of several covariates of the player. Note that, Geometric distribution is a special case Negative Binomial distribution when the number of failures is 1.