Set minimum Julia version to 1.6 (#514)
* Set minimum Julia version to 1.6

Currently the minimum version is 1.0, which is old and unsupported and
makes development and addition of new features more difficult.

* Fix nightly CI

In Julia 1.10, at-test_broken will error if the expression doesn't
evaluate to a boolean value, making it consistent with at-test. We
actually can fit the model here; the test passes with prior Julia
versions only because the expression doesn't evaluate to a boolean.

(cherry picked from commit 57669ef89de3588ffdbf757d82fd28d5c71f5177)