Adjust tolerance when computing rank of R matrix
When fitting based on a QRPivoted factorization, the tolerance used
for determining the rank of R should reflect the size of the matrix
of which the QR factorization was computed. Hence, a custom rtol
based on that first dimension of X is passed to rank.

For consistency, all rank determinations are not handled via rank_linpred
as the LinPred object has all the relevant information.