Stop considering that Geometric and Negative Binomial have a dispersion parameter
This is just incorrect. Fix tests to reflect correct degrees of freedom
and therefore AIC, BIC and standard errors.

This requires adapting how `negbin` works as we need to count one additional
degree of freedom due to the shape parameter (distinct from the dispersion,
which is 1) being estimated from the data. Do this by wrapping the shape
parameter in an `Estimated` type that we use to detect that degrees of
freedom need adjusting.

Also avoid defining `dispersion_parameter` fallback for any distribution:
it's clearer for users to get a `MethodError` so that they have to define the
appropriate method themselves, than silently returning a wrong value.