Stop considering that Negative Binomial has a dispersion parameter
This is incorrect when the shape parameter has been fixed manually and the
model estimated via `glm`. Fix tests to reflect correct degrees of freedom
and therefore AIC, BIC and standard errors.

This requires adapting how `negbin` works as we need to count one additional
degree of freedom due to the shape parameter (distinct from the dispersion,
which is 1) being estimated from the data. Do this by returning a
`NegativeBinomialModel` instead of a standard GLM.