Adjust ftest
This commit changes the degrees of freedom shown in the table from
the number of estimated parameters (dof) to be the number of degrees
of freedom for the model (residual_dof). This makes it possible to
easily calculate the F-test statistic from the other quantities in
the table with the usual formulas.

Since nobs now returns floats, residual_dof also generally returns
floats, so the degrees of freedom parameters are now stored as floats
instead of integers.

The commit removes the R-squared quantities. They are not needed
and I don't think they add value.

The commit also removes numbering of the rows in the output. I don't
think it is helpful.

The show method is now for MIME"text/plain" since that is more
appropriate for a "decorated" show method that span multiple lines.
In most cases this won't be visible to users.