Adjust show method for fitted models
Only show the coefficients instead of the full coefficients table.
At the time of the fitting, the user cannot pass arguments related
to inference such as the level of the confidence interval, so it
is better to only show this output when calling coeftable explicitly.

Also, include the distribution family and link function when fitting
glms as well as include the number of observations, the residual degrees
of freedom and the residual deviance in the output.

Adjust the show method to include MIME"text/plain" as it is producing
"decorated" output. Finally, explicitly test the show method agaist
reference output instead of just writing it to an IOBuffer.